+++ docs/esbuild.config.js
@@ -27,11 +27,13 @@ const outputFolder = "output"
 // const esbuildOptions = { publicPath: "/my_subfolder/_bridgetown/static" }
 // ```
 const esbuildOptions = {
+  target: "es2020",
   entryPoints: [
     "frontend/javascript/index.js",
     "frontend/javascript/defer.js",
     "frontend/javascript/entrypoints/character-counter.js",
-    "frontend/javascript/entrypoints/starter-kit-setup.js"
+    "frontend/javascript/entrypoints/starter-kit-setup.js",
+    "frontend/javascript/entrypoints/syntax-highlighting.js"
   ],
   splitting: true,
   format: "esm"