+++ tests/docs/specifying-accepted-file-types.test.js
@@ -33,6 +33,9 @@ test("Should not allow non-png files to be added via file chooser", async ({ pag
 test("Should allow PNG files to be added via drag and drop", async ({ page }) => {
   await page.goto(pagePath)
 
+
+  // Need to interact with the page before drag and drop works
+  await page.locator("rhino-editor#png-only").click()
   // Drag and drop
   // https://github.com/microsoft/playwright/issues/13364#issuecomment-1156288428
   // https://github.com/microsoft/playwright/issues/10667#issuecomment-998397241
@@ -44,7 +47,7 @@ test("Should allow PNG files to be added via drag and drop", async ({ page }) =>
     fileType: "image/png",
   });
 
-  await page.dispatchEvent('rhino-editor#png-only .ProseMirror', 'drop', { dataTransfer });
+  await page.dispatchEvent('rhino-editor#png-only', 'drop', { dataTransfer });
   await expect(page.locator("rhino-editor#png-only figure")).toBeVisible()
 })